[MOVE/#1] 리패키징
