+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.NOWSOPTAndroid"
         tools:targetApi="31">
         <activity
-            android:name=".auth.signin.SignInActivity"
+            android:name=".presentation.ui.auth.signin.SignInActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -21,11 +21,11 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".auth.signup.SignUpActivity"
+        <activity android:name=".presentation.ui.auth.signup.SignUpActivity"
             android:exported="false">
 
         </activity>
-        <activity android:name=".MainActivity"
+        <activity android:name=".presentation.ui.MainActivity"
             android:exported="false">
 
         </activity>
