+++ app/src/main/java/com/sopt/now/presentation/utils/IntentExt.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.utils
+package com.sopt.now.presentation.utils
 
 import android.content.Intent
 import android.os.Build
