+++ app/src/main/res/layout/activity_signin.xml
@@ -5,7 +5,7 @@
     android:id="@+id/signInActivity"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".presentation.ui.MainActivity">
 
     <TextView
         android:id="@+id/tvSignInTitle"
